Creating system test helper
