+++ test/accessibility_test.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require "application_system_test_case"
+require "system/test_case"
 
-class AccessibilityTest < ApplicationSystemTestCase
+class AccessibilityTest < System::TestCase
   # Skip components that should be tested as part of a larger component.
   # Do not add to this list for any other reason!
   IGNORED_PREVIEWS = %w[
