+++ test/system/test_helper.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require "test_helper"
+
+ENV["TZ"] = "Asia/Taipei"
