+++ test/test_helper.rb
@@ -20,9 +20,8 @@ require "rails/test_help"
 require "view_component/test_helpers"
 require "test_helpers/assert_allocations_helper"
 require "test_helpers/component_test_helpers"
-require "yard/docs_helper"
-require "pry"
-require "yaml"
+# require "pry"
+# require "yaml"
 require "action_controller/railtie"
 require "rails/test_unit/railtie"
 require "active_model/railtie"
